Main Functionality Almost Completed
User is presented with a screen which displays all saved places.
They can filter these saved places with the segmented control as desired.
By clicking the '+' button, the user navigates to a separate VC where they
can search for new places.  After typing in the search bar and clicking
'search', we utilize the Google Places API to fetch places by the entered
text.  Search results are presented in the table view. User can click on
table view cells to reveal more info about the returned place and also
click the 'add' button to save the place.  Any saved places are passed
back to the SavedPlacesViewController through delegation to be displayed.

To Do
- Utilize Google Place Photo API to download photos of each place
- Clean up / Improve UI
- Persist SavedPlaces through CoreData

Notes
- Google API Key not committed, a project build will fail due to missing file