Google Place Photo API Call & PlaceViewModel
Each place found triggers a Google Place Photo api call using
the returned photo_reference id.  This happens in the background
once the respective PlaceViewModel is initialized. These images
are displayed when a user taps on any results returned from a search.
The PlaceViewModel is now is charge of handling all state changes for
Place model.

To Do
- Persist Models to Core Data
- UI / UX updates
- Get Place's phone number with Google's Place Details API